sellowl-backend/requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
firebase-admin>=6.5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
